Extract benchmark Args and scenarios to schemas/benchmark
Move all benchmark Args classes (profiles, outputs, metrics, random),
BenchmarkScenario, BenchmarkArgs, TransientPhaseConfig, and scenario
JSON files into centralized schemas/benchmark/ module. Original files
become thin re-exports. SchedulerState gated behind TYPE_CHECKING in
TransientPhaseConfig.

Generated-by: Claude Opus 4.6
