+++ src/guidellm/benchmark/outputs/plot.py
@@ -13,9 +13,8 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Literal
 
-from pydantic import Field, field_validator
+from pydantic import Field
 
-from guidellm import settings
 from guidellm.benchmark.outputs.output import GenerativeBenchmarkerOutput
 from guidellm.benchmark.schemas import (
     BenchmarkOutputArgs,
@@ -24,15 +23,16 @@ from guidellm.benchmark.schemas import (
 )
 from guidellm.extras import plot
 from guidellm.schemas import StatusBreakdown
+from guidellm.schemas.benchmark.outputs import PlotBenchmarkOutputArgs
 
 __all__ = [
     "GenerativeBenchmarkerPlot",
+    "PlotBenchmarkOutputArgs",
 ]
 
 
 _StatusName = Literal["successful", "incomplete", "errored", "total"]
 _PlotFunction = Callable[["plot.Axes", Sequence["_BenchmarkPoint"]], None]
-_ALLOWED_PLOT_SUFFIXES = {".png", ".jpg", ".jpeg", ".svg", ".pdf"}
 
 
 _BLUE = "#0077b6"
@@ -42,46 +42,6 @@ _GREEN = "#2d6a4f"
 _ORANGE = "#f77f00"
 _PINK = "#c9184a"
 _GRAY = "#6c757d"
-
-
-@BenchmarkOutputArgs.register("plot")
-class PlotBenchmarkOutputArgs(BenchmarkOutputArgs):
-    """Model for Plot benchmark output arguments.
-
-    Defines parameters for generating static image visualizations, enforcing
-    image output suffix.
-    """
-
-    kind: Literal["plot"] = Field(
-        default="plot",
-        description="Type identifier for the plot configuration.",
-    )
-    path: Path = Field(
-        default_factory=lambda: settings.default_results_dir / "benchmarks.png",
-        description="The file to save the output plot to.",
-    )
-    dpi: int = Field(
-        default=100,
-        description="Resolution of the output image in Dots Per Inch.",
-    )
-
-    @field_validator("path", mode="after")
-    @classmethod
-    def validate_plot_suffix(cls, v: Path) -> Path:
-        """Ensures the output file path ends with a supported plotting format extension.
-
-        If the suffix is missing, it defaults to .png.
-        If an unsupported suffix is provided, it raises a ValueError.
-        """
-        if not v.suffix:
-            return v.with_suffix(".png")
-        suffix = v.suffix.lower()
-        if suffix in _ALLOWED_PLOT_SUFFIXES:
-            return v
-        raise ValueError(
-            f"Plot output type {suffix} is not supported: valid types are "
-            f"{', '.join(sorted(_ALLOWED_PLOT_SUFFIXES))}"
-        )
 
 
 @dataclass(frozen=True)
