+++ src/guidellm/benchmark/outputs/serialized.py
@@ -16,45 +16,16 @@ from pydantic import Field
 
 from guidellm.benchmark.outputs.output import GenerativeBenchmarkerOutput
 from guidellm.benchmark.schemas import BenchmarkOutputArgs, GenerativeBenchmarksReport
-from guidellm.settings import settings
+from guidellm.schemas.benchmark.outputs import (
+    JSONBenchmarkOutputArgs,
+    YAMLBenchmarkOutputArgs,
+)
 
 __all__ = [
     "GenerativeBenchmarkerSerialized",
     "JSONBenchmarkOutputArgs",
     "YAMLBenchmarkOutputArgs",
 ]
-
-
-@BenchmarkOutputArgs.register("json")
-class JSONBenchmarkOutputArgs(BenchmarkOutputArgs):
-    """Model for JSON benchmark output arguments."""
-
-    kind: Literal["json"] = Field(
-        default="json",
-        description="The kind of output.",
-        examples=["json"],
-    )
-    path: Path = Field(
-        default_factory=lambda: settings.default_results_dir / "benchmarks.json",
-        description="The file to save the output to.",
-        examples=["./benchmarks.json"],
-    )
-
-
-@BenchmarkOutputArgs.register("yaml")
-class YAMLBenchmarkOutputArgs(BenchmarkOutputArgs):
-    """Model for YAML benchmark output arguments."""
-
-    kind: Literal["yaml"] = Field(
-        default="yaml",
-        description="The kind of output.",
-        examples=["yaml"],
-    )
-    path: Path = Field(
-        default_factory=lambda: settings.default_results_dir / "benchmarks.yaml",
-        description="The file to save the output to.",
-        examples=["./benchmarks.yaml"],
-    )
 
 
 @GenerativeBenchmarkerOutput.register(["json", "yaml"])
