+++ src/guidellm/benchmark/profiles/asynchronous.py
@@ -2,67 +2,23 @@
 
 from __future__ import annotations
 
-import contextlib
 from collections.abc import MutableMapping
 from typing import TYPE_CHECKING, Any, Literal
 
-from pydantic import Field, PositiveFloat, PositiveInt, field_validator
-
-from guidellm.benchmark.schemas import ProfileArgs
 from guidellm.scheduler import (
     AsyncConstantStrategy,
     AsyncPoissonStrategy,
     ConstraintInitializer,
     SchedulingStrategy,
 )
-from guidellm.utils.imports import json
+from guidellm.schemas.benchmark.profiles import AsyncProfileArgs
 
 from .profile import Profile, ProfileFactory
 
+__all__ = ["AsyncProfile", "AsyncProfileArgs"]
+
 if TYPE_CHECKING:
     from guidellm.benchmark.schemas import Benchmark
-
-
-@ProfileArgs.register(["async", "constant", "poisson"])
-class AsyncProfileArgs(ProfileArgs):
-    """Pydantic model for asynchronous profile creation arguments."""
-
-    kind: Literal["async", "constant", "poisson"] = Field(
-        default="async",
-        description="Profile type discriminator for asynchronous scheduling",
-    )
-    rate: list[PositiveFloat] = Field(
-        description="Request scheduling rates in requests per second",
-        examples=[1.0, [1.0, 2.0, 3.0]],
-    )
-    max_concurrency: PositiveInt | None = Field(
-        default=None,
-        description="Maximum concurrent requests to schedule",
-        examples=[10],
-    )
-
-    @field_validator("rate", mode="before")
-    @classmethod
-    def _coerce_rate_to_list(
-        cls, value: list[PositiveFloat] | PositiveFloat
-    ) -> list[PositiveFloat]:
-        """Normalize rate to a list of integers.
-
-        Allow single integer or list of integers.
-        """
-        if isinstance(value, str):
-            with contextlib.suppress(json.JSONDecodeError, ValueError):
-                value = json.loads(value)
-        if not value:
-            raise ValueError("rate requires at least one value")
-        if isinstance(value, list | tuple):
-            return value
-        if isinstance(value, int | float):
-            return [value]
-        raise ValueError(
-            "rate must be a number or a list of numeric values, "
-            f"got {type(value).__name__}"
-        )
 
 
 @ProfileFactory.register(["async", "constant", "poisson"])
