+++ src/guidellm/schemas/benchmark/outputs/__init__.py
@@ -1,0 +1,19 @@
+from guidellm.schemas.benchmark.outputs.console import ConsoleBenchmarkOutputArgs
+from guidellm.schemas.benchmark.outputs.csv import CSVBenchmarkOutputArgs
+from guidellm.schemas.benchmark.outputs.html import HTMLBenchmarkOutputArgs
+from guidellm.schemas.benchmark.outputs.output import BenchmarkOutputArgs
+from guidellm.schemas.benchmark.outputs.plot import PlotBenchmarkOutputArgs
+from guidellm.schemas.benchmark.outputs.serialized import (
+    JSONBenchmarkOutputArgs,
+    YAMLBenchmarkOutputArgs,
+)
+
+__all__ = [
+    "BenchmarkOutputArgs",
+    "CSVBenchmarkOutputArgs",
+    "ConsoleBenchmarkOutputArgs",
+    "HTMLBenchmarkOutputArgs",
+    "JSONBenchmarkOutputArgs",
+    "PlotBenchmarkOutputArgs",
+    "YAMLBenchmarkOutputArgs",
+]
