+++ src/guidellm/schemas/benchmark/profiles/__init__.py
@@ -1,0 +1,17 @@
+from guidellm.schemas.benchmark.profiles.asynchronous import AsyncProfileArgs
+from guidellm.schemas.benchmark.profiles.concurrent import ConcurrentProfileArgs
+from guidellm.schemas.benchmark.profiles.profile import ProfileArgs
+from guidellm.schemas.benchmark.profiles.replay import ReplayProfileArgs
+from guidellm.schemas.benchmark.profiles.sweep import SweepProfileArgs
+from guidellm.schemas.benchmark.profiles.synchronous import SynchronousProfileArgs
+from guidellm.schemas.benchmark.profiles.throughput import ThroughputProfileArgs
+
+__all__ = [
+    "AsyncProfileArgs",
+    "ConcurrentProfileArgs",
+    "ProfileArgs",
+    "ReplayProfileArgs",
+    "SweepProfileArgs",
+    "SynchronousProfileArgs",
+    "ThroughputProfileArgs",
+]
